refactor logging out of ina219 driver
Logging should be configured by parent modules. Although the code supported this, there should be no need to have an internal configuration of logging.

Removing this dependency makes logging cleaner.

The API of INA219 is kept compatible to older versions, but the `log_level` arg is now ignored and marked as deprecated.